Use scaffolding to generate files for the game resource
Generated with `mix phx.gen.html User Game game`.

Then ` resources "/games", GameController` was added to
`lib/missiles_web/router.ex`. It needs to be in the browser scope. I
think the `pipe_through :browser` on line 18 does that, even though it
is in the scope '/' block.

After all that `mix ecto.migrate` created the database table.

For the generator the syntax is CONTEXT (User) then singular and plural
forms of the resource (Game game). I think what this means is the game
in the context of the user playing it.