Make sure the ships attributes are saved to the database
This required specifying the fields that need to be case and validated
from the parameters are listed in the `cast` and `validate_required`
functions.

The index view now displays ships correctly. The ship form can now take
values for input. The show page will show a ships details.

There are now a whole bunch of ship records in the database with no name
or category values, so deleting an existing ship might be a good next
step, or editing an existing ship.