Change User context to Games
Like the change of Fleet context to Ships I am naming the contexts as
the plural of the resource, following the advice of the Contexts
documentation.

What I want from the Game resource is to be the main point of
interaction between the user and the app. The `show` action of a Game
will be the place where a Game is 'played' by the user.