Update the 'turn' counter in a game
Since the 'show' view is being used as a game UI it needs a form to take
updates to the 'turn' field. It's unconventional to edit state in the
'show' action, but it works for now.

An ideal next feature might be to allow editing of the 'turn' field at
the game 'edit' view, so I can reset the field during development.